pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vuln-risk-assessment"
version = "0.1.0"
description = "Hybrid AMM + ML Vulnerability Risk Assessment System"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "scipy>=1.11",
    "scikit-learn>=1.3",
    "xgboost>=2.0",
    "networkx>=3.1",
    "duckdb>=0.9",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "fastapi>=0.104",
    "uvicorn>=0.24",
    "strawberry-graphql>=0.215",
    "joblib>=1.3",
    "pandas>=2.1",
]

[project.optional-dependencies]
embeddings = [
    "sentence-transformers>=2.2",
    "torch>=2.1",
]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
